refactor(agent): satisfy repolint ratchet and golangci for the 400 repair
Zero behavior change: lint-gate compliance for the thinking-replay repair
branch. The branch's additions pushed four files past their repolint budgets
and introduced two golangci findings; everything here is code motion or
mechanical rewrites, no logic edits.

repolint:
- run_loop.go file-size/function-size: deferredStreamSink and its methods move
  verbatim to the new stream_sink.go; the thinking-400 branch collapses into
  tryRecoverReasoningReplay400 in reasoning_replay_recovery.go (Discard,
  attempt audit, and budget handling live there now); the overflow and
  suppressed-retry tails of streamWithSamplingRecovery move verbatim into
  finishReasoningReplayOverflow / suppressMissingReasoningRetry in
  reasoning_replay.go.
- sessionstate.go essay: the reasoningReplayStrongProjection field comment
  drops to three lines.
- anthropic.go file-size: the thinking-first merge logic moves to
  missing_reasoning_fallback.go as mergeThinkingFirst + leadingThinkingBlocks,
  leaving appendBlocks a one-line call.
- anthropic_test.go test-file-size: TestBuildRequestDeepSeekReplaysReasoning-
  FromHistory moves unchanged to replay_projection_test.go, the file that owns
  replay/projection semantics.

golangci-lint:
- errorlint in reasoning_replay_error_test.go: assert the Unwrap contract via
  errors.Is instead of comparing errors.Unwrap results.
- modernize in retry_e2e_test.go: accumulate the streamed answer with a
  strings.Builder.

Verification: go run ./tools/repolint (clean); golangci-lint run
--timeout=5m ./internal/... (0 issues); go build ./...; go test -count=1
./internal/provider/... ./internal/agent/...; go test -race
./internal/agent/ -run 'Replay|Reasoning|Retry' -count=1.
