+++ internal/agent/run_loop.go
@@ -46,78 +46,6 @@ func (s streamedTurn) assistantMessage() provider.Message {
 	}
 }
 
-// deferredStreamSink keeps selected stream events local until the caller
-// chooses which provider response to adopt. On an ordinary healthy DeepSeek
-// turn, reasoning arrives before tool calls and unlocks live tool-card events.
-// On the rare malformed turn with no reasoning, only the speculative partial
-// tool cards remain buffered, so retrying does not flash duplicate cards in the
-// UI. A recovery attempt buffers everything because it may be discarded.
-type deferredStreamSink struct {
-	inner               event.Sink
-	deferAll            bool
-	waitingForReasoning bool
-	sawReasoning        bool
-	events              []event.Event
-}
-
-func newReasoningAwareStreamSink(inner event.Sink) *deferredStreamSink {
-	return &deferredStreamSink{inner: inner, waitingForReasoning: true}
-}
-
-func newDeferredStreamSink(inner event.Sink) *deferredStreamSink {
-	return &deferredStreamSink{inner: inner, deferAll: true}
-}
-
-func (s *deferredStreamSink) Emit(e event.Event) {
-	if s == nil {
-		return
-	}
-	if s.deferAll {
-		s.events = append(s.events, e)
-		return
-	}
-	if s.waitingForReasoning && e.Kind == event.Reasoning && strings.TrimSpace(e.Text) != "" {
-		s.sawReasoning = true
-		s.inner.Emit(e)
-		s.flushBuffered()
-		return
-	}
-	if s.waitingForReasoning && !s.sawReasoning {
-		switch e.Kind {
-		case event.ToolDispatch, event.ToolResult, event.Text, event.Message:
-			// Keep every user-visible speculative event private until reasoning
-			// proves the turn replayable. Healthy DeepSeek responses emit
-			// reasoning first, so their live-streaming fast path is unchanged.
-			s.events = append(s.events, e)
-			return
-		}
-	}
-	s.inner.Emit(e)
-}
-
-func (s *deferredStreamSink) flushBuffered() {
-	if s == nil {
-		return
-	}
-	for _, e := range s.events {
-		s.inner.Emit(e)
-	}
-	s.events = nil
-}
-
-func (s *deferredStreamSink) Flush() {
-	if s == nil {
-		return
-	}
-	s.flushBuffered()
-}
-
-func (s *deferredStreamSink) Discard() {
-	if s != nil {
-		s.events = nil
-	}
-}
-
 // beginRunTurn handles evidence scope, delivery classification, background-job
 // evidence re-lease, and the initial user-turn persistence. Callers still own
 // all Run-level defers (workspace lease, evidence commit, delivery checkpoint,
@@ -392,15 +320,7 @@ func (a *Agent) streamWithSamplingRecovery(ctx context.Context, turn int) stream
 				attempt = 0
 				continue
 			}
-			// A thinking-400 means the frozen history's replayed reasoning is
-			// stale for this provider. Repair the projection once and retry;
-			// every other 400 keeps falling through to the terminal path.
-			if next, retryReplay := a.recoverReasoningReplay400(frozen, result.err, &replayRecovery); retryReplay {
-				if streamSink != nil {
-					streamSink.Discard()
-				}
-				a.emitStreamAttempt(attemptID, event.StreamAttemptDiscard, attempt, "reasoning_replay_400", result.err)
-				event.RecordProtocolRecovery(a.svc.sink, event.ProtocolRecoveryAudit{Kind: event.ProtocolRecoveryReasoningReplay400Detected})
+			if next, retryReplay := a.tryRecoverReasoningReplay400(streamSink, frozen, attemptID, attempt, result.err, &replayRecovery); retryReplay {
 				frozen = next
 				reasoningReplayRepaired = true
 				attempt = 0
@@ -431,11 +351,7 @@ func (a *Agent) streamWithSamplingRecovery(ctx context.Context, turn int) stream
 			a.observeMissingAssistantReasoning(result.assistantMessage(), result.reasoningComplete)
 		}
 		if issue == ReasoningReplayOverflow {
-			event.RecordProtocolRecovery(a.svc.sink, event.ProtocolRecoveryAudit{Kind: event.ProtocolRecoveryReasoningOverflowDetected})
-			result.usage = finalizeSamplingUsage(billable, result.usage)
-			terminal := a.finishUnreplayableReasoning(result, streamSink, issue)
-			a.emitReasoningReplayAttemptOutcome(attemptID, attempt, terminal.err)
-			return terminal
+			return a.finishReasoningReplayOverflow(result, streamSink, issue, billable, attemptID, attempt)
 		}
 		if missing {
 			event.RecordProtocolRecovery(a.svc.sink, event.ProtocolRecoveryAudit{Kind: event.ProtocolRecoveryMissingReasoningDetected})
@@ -475,15 +391,7 @@ func (a *Agent) streamWithSamplingRecovery(ctx context.Context, turn int) stream
 				return retry
 			}
 			if !shouldRetry {
-				event.RecordProtocolRecovery(a.svc.sink, event.ProtocolRecoveryAudit{Kind: event.ProtocolRecoveryMissingReasoningRetrySuppressed})
-				if fallback, ok := a.runMissingReasoningFallback(ctx, turn, &frozen, attemptID, attempt, billable, streamSink); ok {
-					return fallback
-				}
-				event.RecordProtocolRecovery(a.svc.sink, event.ProtocolRecoveryAudit{Kind: event.ProtocolRecoveryMissingReasoningFallback})
-				result.usage = finalizeSamplingUsage(billable, result.usage)
-				terminal := a.finishUnreplayableReasoning(result, streamSink, issue)
-				a.emitReasoningReplayAttemptOutcome(attemptID, attempt, terminal.err)
-				return terminal
+				return a.suppressMissingReasoningRetry(ctx, turn, &frozen, attemptID, attempt, streamSink, result, issue, billable)
 			}
 		}
 
@@ -491,9 +399,6 @@ func (a *Agent) streamWithSamplingRecovery(ctx context.Context, turn int) stream
 		a.emitStreamAttempt(attemptID, event.StreamAttemptCommit, attempt, "", nil)
 		result.usage = finalizeSamplingUsage(billable, result.usage)
 		if reasoningReplayRepaired {
-			// The repair retry proved the canonical history's stored reasoning is
-			// stale for this provider; keep the strong projection for the rest of
-			// the conversation so later rounds do not pay another 400.
 			a.activateReasoningReplayStrongProjection()
 		}
 		return result
