+++ internal/provider/anthropic/anthropic_test.go
@@ -702,56 +702,6 @@ func TestMissingToolCallReasoningWarningFingerprintTracksAnthropicConfiguration(
 	}
 }
 
-func TestBuildRequestDeepSeekReplaysReasoningFromHistory(t *testing.T) {
-	toolTurn := []provider.Message{
-		{Role: provider.RoleUser, Content: "weather?"},
-		{Role: provider.RoleAssistant, ReasoningContent: "I should call the tool.",
-			ToolCalls: []provider.ToolCall{{ID: "t1", Name: "get_weather", Arguments: `{"city":"Paris"}`}}},
-		{Role: provider.RoleTool, ToolCallID: "t1", Content: "sunny"},
-	}
-	for _, tc := range []struct {
-		name     string
-		thinking string
-		effort   string
-	}{
-		{name: "current request has no tools", thinking: "enabled", effort: "high"},
-		{name: "thinking disabled after tool call", thinking: "enabled", effort: "disabled"},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			c := &client{model: "deepseek-v4-flash", deepseek: true, thinking: tc.thinking, effort: tc.effort}
-			r := c.buildRequest(context.Background(), provider.Request{Messages: toolTurn})
-			if len(r.Tools) != 0 {
-				t.Fatalf("current request tools = %+v, want none", r.Tools)
-			}
-			if len(r.Messages) != 3 {
-				t.Fatalf("messages = %+v, want user/assistant/user", r.Messages)
-			}
-			blocks := r.Messages[1].Content
-			if len(blocks) != 2 || blocks[0].Type != "thinking" || blocks[0].Thinking != "I should call the tool." || blocks[1].Type != "tool_use" {
-				t.Fatalf("assistant blocks = %+v, want historical thinking before tool_use", blocks)
-			}
-		})
-	}
-
-	t.Run("reasoning without a tool call is replayed", func(t *testing.T) {
-		// DeepSeek's thinking mode requires every historical assistant turn's
-		// thinking block back when the request declares tools, even for plain
-		// question-answer turns (api-docs.deepseek.com/guides/thinking_mode).
-		c := &client{model: "deepseek-v4-flash", deepseek: true, thinking: "enabled", effort: "high"}
-		r := c.buildRequest(context.Background(), provider.Request{
-			Messages: []provider.Message{
-				{Role: provider.RoleUser, Content: "hello"},
-				{Role: provider.RoleAssistant, Content: "hi", ReasoningContent: "scratchpad"},
-			},
-			Tools: []provider.ToolSchema{{Name: "get_weather"}},
-		})
-		blocks := r.Messages[1].Content
-		if len(r.Messages) != 2 || len(blocks) != 2 || blocks[0].Type != "thinking" || blocks[0].Thinking != "scratchpad" || blocks[1].Type != "text" {
-			t.Fatalf("non-tool assistant blocks = %+v, want thinking before text", r.Messages)
-		}
-	})
-}
-
 func TestBuildRequestDeepSeekThinkingModes(t *testing.T) {
 	for _, tc := range []struct {
 		name, model, input, want string
