+++ internal/provider/anthropic/missing_reasoning_fallback.go
@@ -37,6 +37,33 @@ func (c *client) replayReasoningBlock(m provider.Message, recoveryWithoutThinkin
 	return contentBlock{}, false
 }
 
+// mergeThinkingFirst merges appended blocks into an existing same-role
+// message's content. A thinking block must lead its message, so a leading
+// thinking run in the appended blocks moves to the front: after projection
+// degrades a reasoning-less turn to plain text, a following healthy assistant
+// turn merging into it must stay [thinking, text, ...], never [text, ...].
+func mergeThinkingFirst(existing, blocks []contentBlock) []contentBlock {
+	head := leadingThinkingBlocks(blocks)
+	if head == 0 || len(existing) == 0 {
+		return append(existing, blocks...)
+	}
+	merged := make([]contentBlock, 0, len(existing)+len(blocks))
+	merged = append(merged, blocks[:head]...)
+	merged = append(merged, existing...)
+	merged = append(merged, blocks[head:]...)
+	return merged
+}
+
+// leadingThinkingBlocks counts the thinking blocks at the head of blocks.
+// replayReasoningBlock emits at most one, so the run is normally 0 or 1.
+func leadingThinkingBlocks(blocks []contentBlock) int {
+	head := 0
+	for head < len(blocks) && blocks[head].Type == "thinking" {
+		head++
+	}
+	return head
+}
+
 func (c *client) applyDeepSeekThinking(r *anthRequest, req provider.Request, recoveryWithoutThinking bool) {
 	r.Temperature = req.Temperature
 	t := c.thinking
